Remove any traces of `BlurHashAsyncImage` in `:features:messages`, since it was moved to the design system.
